@@ -1,84 +1,51 @@
 {
-  "name": "medusa-next",
-  "version": "1.0.3",
+  "name": "storefront",
   "private": true,
-  "author": "Kasper Fabricius Kristensen <kasper@medusajs.com> & Victor Gerbrands <victor@medusajs.com> (https://www.medusajs.com)",
-  "description": "Next.js Starter to be used with Medusa server",
-  "keywords": [
-    "medusa-storefront"
-  ],
+  "type": "module",
   "scripts": {
-    "wait": "await-backend",
-    "launcher": "launch-storefront",
-    "dev": "npm run launcher dev",
-    "build": "npm run launcher build",
-    "start": "npm run launcher start",
-    "build:next": "next build",
-    "lint": "next lint",
-    "analyze": "ANALYZE=true next build",
-    "test-e2e": "playwright test e2e"
-  },
-  "resolutions": {
-    "webpack": "^5"
-  },
-  "pnpm": {
-    "overrides": {
-      "glob@>=10.2.0 <10.5.0": ">=10.5.0",
-      "@babel/runtime@<7.26.10": ">=7.26.10",
-      "@babel/helpers@<7.26.10": ">=7.26.10",
-      "prismjs@<1.30.0": ">=1.30.0",
-      "js-yaml@>=4.0.0 <4.1.1": ">=4.1.1",
-      "brace-expansion@>=1.0.0 <=1.1.11": ">=1.1.12",
-      "brace-expansion@>=2.0.0 <=2.0.1": ">=2.0.2"
-    }
+    "dev": "vite dev",
+    "build": "vite build",
+    "build:prod": "pnpm build && pnpm preview",
+    "preview": "vite preview"
   },
   "dependencies": {
-    "@headlessui/react": "^1.6.1",
-    "@hookform/error-message": "^2.0.0",
-    "@medusajs/js-sdk": "preview",
-    "@medusajs/types": "preview",
-    "@medusajs/ui": "preview",
-    "@meilisearch/instant-meilisearch": "^0.7.1",
-    "@paypal/paypal-js": "^5.0.6",
-    "@paypal/react-paypal-js": "^7.8.1",
-    "@stripe/react-stripe-js": "^1.7.2",
-    "@stripe/stripe-js": "^1.29.0",
-    "algoliasearch": "^4.20.0",
-    "axios": "^1.12.0",
-    "lodash": "^4.17.21",
-    "lucide-react": "^0.562.0",
-    "medusajs-launch-utils": "0.0.18",
-    "next": "^15.5.9",
-    "pg": "^8.11.3",
-    "qs": "^6.12.1",
-    "react": "^18.2.0",
-    "react-country-flag": "^3.0.2",
-    "react-dom": "^18.2.0",
-    "react-instantsearch-hooks-web": "^6.29.0",
-    "react-intersection-observer": "^9.3.4",
-    "server-only": "^0.0.1",
-    "tailwindcss-radix": "^2.8.0",
-    "webpack": "^5"
+    "@medusajs/icons": "2.11.3-preview-20251103060142",
+    "@medusajs/js-sdk": "2.11.3-preview-20251103060142",
+    "@radix-ui/react-dialog": "^1.1.15",
+    "@radix-ui/react-navigation-menu": "^1.2.14",
+    "@radix-ui/react-select": "^2.2.6",
+    "@tanstack/react-query": "5.66.2",
+    "@tanstack/react-router": "1.131.32",
+    "@tanstack/react-router-ssr-query": "1.131.32",
+    "@tanstack/react-start": "1.131.32",
+    "class-variance-authority": "^0.7.1",
+    "clsx": "^2.1.1",
+    "dotenv": "17.2.1",
+    "lodash-es": "4.17.21",
+    "react": "19.1.1",
+    "react-dom": "19.1.1",
+    "tailwindcss": "4.1.12",
+    "tailwindcss-animate": "^1.0.7",
+    "vite-tsconfig-paths": "5.1.4"
   },
   "devDependencies": {
-    "@babel/core": "^7.17.5",
-    "@medusajs/client-types": "preview",
-    "@medusajs/ui-preset": "preview",
-    "@playwright/test": "^1.55.1",
-    "@types/lodash": "^4.14.195",
-    "@types/node": "17.0.21",
-    "@types/pg": "^8.11.0",
-    "@types/react": "^18.2.42",
-    "@types/react-dom": "^18.2.18",
-    "@types/react-instantsearch-dom": "^6.12.3",
-    "ansi-colors": "^4.1.3",
-    "autoprefixer": "^10.4.2",
-    "babel-loader": "^8.2.3",
-    "eslint": "^8.57.0",
-    "eslint-config-next": "^13.4.5",
-    "postcss": "^8.4.8",
-    "prettier": "^2.8.8",
-    "tailwindcss": "^3.0.23",
-    "typescript": "^5.3.2"
+    "@eslint/js": "9.33.0",
+    "@medusajs-ai/tags": "next",
+    "@medusajs/types": "2.11.3-preview-20251103060142",
+    "@tailwindcss/vite": "4.1.4",
+    "@tanstack/react-router-devtools": "1.131.32",
+    "@types/lodash-es": "4.17.12",
+    "@types/react": "19.1.1",
+    "@types/react-dom": "19.1.1",
+    "@vitejs/plugin-react": "5.0.1",
+    "eslint": "9.33.0",
+    "eslint-import-resolver-typescript": "4.4.4",
+    "eslint-plugin-react-hooks": "5.2.0",
+    "eslint-plugin-react-refresh": "0.4.20",
+    "globals": "16.3.0",
+    "typescript": "5.8.3",
+    "typescript-eslint": "8.39.1",
+    "vite": "7.1.2",
+    "vite-plugin-terminal": "1.3.0"
   }
 }
