@@ -1,398 +1,365 @@
-"use server"
-
-import { sdk } from "@lib/config"
-import medusaError from "@lib/util/medusa-error"
+import { sdk } from "@/lib/utils/sdk"
+import { getRegion } from "@/lib/data/regions"
+import { getStoredCart, setStoredCart } from "@/lib/utils/cart"
 import { HttpTypes } from "@medusajs/types"
-import { omit } from "lodash"
-import { revalidateTag } from "next/cache"
-import { redirect } from "next/navigation"
-import { getAuthHeaders, getCartId, removeCartId, setCartId } from "./cookies"
-import { getProductsById } from "./products"
-import { getRegion } from "./regions"
+import { 
+  sendPostRequest, 
+  sendDeleteRequest 
+} from "@/lib/data/custom"
 
-export async function retrieveCart() {
-  const cartId = await getCartId()
+const DEFAULT_CART_FIELDS = "+items.total, shipping_methods.name"
 
-  if (!cartId) {
+/**
+ * Retrieves a cart by ID or from stored ID. Returns null if no cart is found.
+ * 
+ * @param cart_id - Optional cart ID. If not provided, will use stored cart ID.
+ * @param fields - Optional fields to include in the response
+ * @returns Promise that resolves to the cart data or null if not found
+ * 
+ * @example
+ * ```typescript
+ * // Get current cart from cookies
+ * const cart = await retrieveCart();
+ * 
+ * // Get specific cart by ID
+ * const specificCart = await retrieveCart({ 
+ *   cart_id: 'cart_123',
+ *   fields: '*items, *items.variant, *items.variant.product'
+ * });
+ * 
+ * // Get cart with minimal fields
+ * const minimalCart = await retrieveCart({ 
+ *   fields: 'id, total, items.quantity'
+ * });
+ * ```
+ */
+export const retrieveCart = async ({
+  cart_id,
+  fields = DEFAULT_CART_FIELDS,
+}: {
+  cart_id?: string;
+  fields?: string;
+}): Promise<HttpTypes.StoreCart | null> => {
+  const id = cart_id || getStoredCart()
+
+  if (!id) {
     return null
   }
 
-  return await sdk.store.cart
-    .retrieve(cartId, {}, { next: { tags: ["cart"] }, ...(await getAuthHeaders()) })
-    .then(({ cart }) => cart)
-    .catch(() => {
-      return null
-    })
-}
-
-export async function getOrSetCart(countryCode: string) {
-  let cart = await retrieveCart()
-  const region = await getRegion(countryCode)
-
-  if (!region) {
-    throw new Error(`Region not found for country code: ${countryCode}`)
-  }
-
-  if (!cart) {
-    const cartResp = await sdk.store.cart.create({ region_id: region.id })
-    cart = cartResp.cart
-    await setCartId(cart.id)
-    revalidateTag("cart")
-  }
-
-  if (cart && cart?.region_id !== region.id) {
-    await sdk.store.cart.update(
-      cart.id,
-      { region_id: region.id },
-      {},
-      await getAuthHeaders()
-    )
-    revalidateTag("cart")
-  }
+  const { cart } = await sdk.store.cart.retrieve(id, {
+    fields:
+      fields,
+  })
 
   return cart
 }
 
-export async function updateCart(data: HttpTypes.StoreUpdateCart) {
-  const cartId = await getCartId()
-  if (!cartId) {
-    throw new Error("No existing cart found, please create one before updating")
-  }
-
-  return sdk.store.cart
-    .update(cartId, data, {}, await getAuthHeaders())
-    .then(({ cart }) => {
-      revalidateTag("cart")
-      return cart
-    })
-    .catch(medusaError)
-}
-
-export async function addToCart({
-  variantId,
-  quantity,
-  countryCode,
+/**
+ * Creates a new cart for the specified region and stores the cart ID.
+ * 
+ * @param region_id - The region ID to create the cart for
+ * @param fields - Optional fields to include in the response (defaults to items total and shipping methods)
+ * @returns Promise that resolves to the newly created cart
+ * 
+ * @example
+ * ```typescript
+ * // Create cart for US region
+ * const newCart = await createCart({ 
+ *   region_id: 'reg_us',
+ *   fields: '*items, *items.variant, shipping_methods'
+ * });
+ * 
+ * // Create cart with minimal response
+ * const cart = await createCart({ 
+ *   region_id: 'reg_eu',
+ *   fields: 'id, total'
+ * });
+ * ```
+ */
+export const createCart = async ({
+  region_id,
+  fields = DEFAULT_CART_FIELDS,
 }: {
-  variantId: string
-  quantity: number
-  countryCode: string
-}) {
-  if (!variantId) {
-    throw new Error("Missing variant ID when adding to cart")
-  }
-
-  const cart = await getOrSetCart(countryCode)
-  if (!cart) {
-    throw new Error("Error retrieving or creating cart")
-  }
-
-  await sdk.store.cart
-    .createLineItem(
-      cart.id,
-      {
-        variant_id: variantId,
-        quantity,
-      },
-      {},
-      await getAuthHeaders()
-    )
-    .then(() => {
-      revalidateTag("cart")
-    })
-    .catch(medusaError)
-}
-
-export async function updateLineItem({
-  lineId,
-  quantity,
-}: {
-  lineId: string
-  quantity: number
-}) {
-  if (!lineId) {
-    throw new Error("Missing lineItem ID when updating line item")
-  }
-
-  const cartId = await getCartId()
-  if (!cartId) {
-    throw new Error("Missing cart ID when updating line item")
-  }
-
-  await sdk.store.cart
-    .updateLineItem(cartId, lineId, { quantity }, {}, await getAuthHeaders())
-    .then(() => {
-      revalidateTag("cart")
-    })
-    .catch(medusaError)
-}
-
-export async function deleteLineItem(lineId: string) {
-  if (!lineId) {
-    throw new Error("Missing lineItem ID when deleting line item")
-  }
-
-  const cartId = await getCartId()
-  if (!cartId) {
-    throw new Error("Missing cart ID when deleting line item")
-  }
-
-  await sdk.store.cart
-    .deleteLineItem(cartId, lineId, {}, await getAuthHeaders())
-    .then(() => {
-      revalidateTag("cart")
-    })
-    .catch(medusaError)
-  revalidateTag("cart")
-}
-
-export async function enrichLineItems(
-  lineItems:
-    | HttpTypes.StoreCartLineItem[]
-    | HttpTypes.StoreOrderLineItem[]
-    | null,
-  regionId: string
-) {
-  if (!lineItems) return []
-
-  // Prepare query parameters
-  const queryParams = {
-    ids: lineItems.map((lineItem) => lineItem.product_id!),
-    regionId: regionId,
-  }
-
-  // Fetch products by their IDs
-  const products = await getProductsById(queryParams)
-  // If there are no line items or products, return an empty array
-  if (!lineItems?.length || !products) {
-    return []
-  }
-
-  // Enrich line items with product and variant information
-  const enrichedItems = lineItems.map((item) => {
-    const product = products.find((p: any) => p.id === item.product_id)
-    const variant = product?.variants?.find(
-      (v: any) => v.id === item.variant_id
-    )
-
-    // If product or variant is not found, return the original item
-    if (!product || !variant) {
-      return item
-    }
-
-    // If product and variant are found, enrich the item
-    return {
-      ...item,
-      variant: {
-        ...variant,
-        product: omit(product, "variants"),
-      },
-    }
-  }) as HttpTypes.StoreCartLineItem[]
-
-  return enrichedItems
-}
-
-export async function setShippingMethod({
-  cartId,
-  shippingMethodId,
-}: {
-  cartId: string
-  shippingMethodId: string
-}) {
-  return sdk.store.cart
-    .addShippingMethod(
-      cartId,
-      { option_id: shippingMethodId },
-      {},
-      await getAuthHeaders()
-    )
-    .then(() => {
-      revalidateTag("cart")
-    })
-    .catch(medusaError)
-}
-
-export async function initiatePaymentSession(
-  cart: HttpTypes.StoreCart,
-  data: {
-    provider_id: string
-    context?: Record<string, unknown>
-  }
-) {
-  return sdk.store.payment
-    .initiatePaymentSession(cart, data, {}, await getAuthHeaders())
-    .then((resp) => {
-      revalidateTag("cart")
-      return resp
-    })
-    .catch(medusaError)
-}
-
-export async function applyPromotions(codes: string[]) {
-  const cartId = await getCartId()
-  if (!cartId) {
-    throw new Error("No existing cart found")
-  }
-
-  await updateCart({ promo_codes: codes })
-    .then(() => {
-      revalidateTag("cart")
-    })
-    .catch(medusaError)
-}
-
-export async function applyGiftCard(code: string) {
-  //   const cartId = getCartId()
-  //   if (!cartId) return "No cartId cookie found"
-  //   try {
-  //     await updateCart(cartId, { gift_cards: [{ code }] }).then(() => {
-  //       revalidateTag("cart")
-  //     })
-  //   } catch (error: any) {
-  //     throw error
-  //   }
-}
-
-export async function removeDiscount(code: string) {
-  // const cartId = getCartId()
-  // if (!cartId) return "No cartId cookie found"
-  // try {
-  //   await deleteDiscount(cartId, code)
-  //   revalidateTag("cart")
-  // } catch (error: any) {
-  //   throw error
-  // }
-}
-
-export async function removeGiftCard(
-  codeToRemove: string,
-  giftCards: any[]
-  // giftCards: GiftCard[]
-) {
-  //   const cartId = getCartId()
-  //   if (!cartId) return "No cartId cookie found"
-  //   try {
-  //     await updateCart(cartId, {
-  //       gift_cards: [...giftCards]
-  //         .filter((gc) => gc.code !== codeToRemove)
-  //         .map((gc) => ({ code: gc.code })),
-  //     }).then(() => {
-  //       revalidateTag("cart")
-  //     })
-  //   } catch (error: any) {
-  //     throw error
-  //   }
-}
-
-export async function submitPromotionForm(
-  currentState: unknown,
-  formData: FormData
-) {
-  const code = formData.get("code") as string
-  try {
-    await applyPromotions([code])
-  } catch (e: any) {
-    return e.message
-  }
-}
-
-// TODO: Pass a POJO instead of a form entity here
-export async function setAddresses(currentState: unknown, formData: FormData) {
-  try {
-    if (!formData) {
-      throw new Error("No form data found when setting addresses")
-    }
-    const cartId = await getCartId()
-    if (!cartId) {
-      throw new Error("No existing cart found when setting addresses")
-    }
-
-    const data = {
-      shipping_address: {
-        first_name: formData.get("shipping_address.first_name"),
-        last_name: formData.get("shipping_address.last_name"),
-        address_1: formData.get("shipping_address.address_1"),
-        address_2: "",
-        company: formData.get("shipping_address.company"),
-        postal_code: formData.get("shipping_address.postal_code"),
-        city: formData.get("shipping_address.city"),
-        country_code: formData.get("shipping_address.country_code"),
-        province: formData.get("shipping_address.province"),
-        phone: formData.get("shipping_address.phone"),
-      },
-      email: formData.get("email"),
-    } as any
-
-    const sameAsBilling = formData.get("same_as_billing")
-    if (sameAsBilling === "on") data.billing_address = data.shipping_address
-
-    if (sameAsBilling !== "on")
-      data.billing_address = {
-        first_name: formData.get("billing_address.first_name"),
-        last_name: formData.get("billing_address.last_name"),
-        address_1: formData.get("billing_address.address_1"),
-        address_2: "",
-        company: formData.get("billing_address.company"),
-        postal_code: formData.get("billing_address.postal_code"),
-        city: formData.get("billing_address.city"),
-        country_code: formData.get("billing_address.country_code"),
-        province: formData.get("billing_address.province"),
-        phone: formData.get("billing_address.phone"),
-      }
-    await updateCart(data)
-  } catch (e: any) {
-    return e.message
-  }
-
-  redirect(
-    `/${formData.get("shipping_address.country_code")}/checkout?step=delivery`
-  )
-}
-
-export async function placeOrder() {
-  const cartId = await getCartId()
-  if (!cartId) {
-    throw new Error("No existing cart found when placing an order")
-  }
-
-  const cartRes = await sdk.store.cart
-    .complete(cartId, {}, await getAuthHeaders())
-    .then((cartRes) => {
-      revalidateTag("cart")
-      return cartRes
-    })
-    .catch(medusaError)
-
-  if (cartRes?.type === "order") {
-    const countryCode =
-      cartRes.order.shipping_address?.country_code?.toLowerCase()
-    await removeCartId()
-    redirect(`/${countryCode}/order/confirmed/${cartRes?.order.id}`)
-  }
-
-  return cartRes.cart
+  region_id: string;
+  fields?: string;
+}): Promise<HttpTypes.StoreCart> => {
+  const { cart } = await sdk.store.cart.create({ region_id }, {
+    fields,
+  })
+  setStoredCart(cart.id)
+  return cart
 }
 
 /**
- * Updates the countrycode param and revalidates the regions cache
- * @param regionId
- * @param countryCode
+ * Updates the current cart with the provided updates.
+ * 
+ * @param updates - The updates to apply to the cart
+ * @param fields - Optional fields to include in the response
+ * @returns Promise that resolves to the updated cart
+ * 
+ * @example
+ * ```typescript
+ * // Update cart with region
+ * const updatedCart = await updateCart({
+ *   region_id: 'reg_us'
+ * });
+ * ```
  */
-export async function updateRegion(countryCode: string, currentPath: string) {
-  const cartId = await getCartId()
-  const region = await getRegion(countryCode)
+export const updateCart = async ({
+  fields = DEFAULT_CART_FIELDS,
+  ...updates
+}: HttpTypes.StoreUpdateCart & {
+  fields?: string;
+}): Promise<HttpTypes.StoreCart> => {
+  const cartId = getStoredCart()
+  if (!cartId) {
+    throw new Error("No cart found")
+  }
+  const { cart } = await sdk.store.cart.update(cartId, updates, {
+    fields,
+  })
+  return cart
+}
 
-  if (!region) {
-    throw new Error(`Region not found for country code: ${countryCode}`)
+/**
+ * Adds a product variant to the cart. Creates a new cart if none exists.
+ * 
+ * @param variant_id - The product variant ID to add to the cart
+ * @param quantity - The quantity of the variant to add
+ * @param country_code - The country code to determine the region (used if creating new cart)
+ * @param fields - Optional fields to include in the response
+ * @returns Promise that resolves to the updated cart
+ * 
+ * @example
+ * ```typescript
+ * // Add single item to cart
+ * const updatedCart = await addToCart({
+ *   variant_id: 'variant_123',
+ *   quantity: 2,
+ *   country_code: 'us',
+ *   fields: '*items, *items.variant, *items.variant.product'
+ * });
+ * 
+ * // Add item with minimal response
+ * const cart = await addToCart({
+ *   variant_id: 'variant_456',
+ *   quantity: 1,
+ *   country_code: 'gb',
+ *   fields: 'id, total, items.quantity'
+ * });
+ * ```
+ */
+export const addToCart = async ({
+  variant_id,
+  quantity,
+  country_code,
+  fields = DEFAULT_CART_FIELDS,
+}: {
+  variant_id: string;
+  quantity: number;
+  country_code: string;
+  fields?: string;
+}): Promise<HttpTypes.StoreCart> => {
+  if (!variant_id) {
+    throw new Error("Missing variant ID when adding to cart")
   }
 
-  if (cartId) {
-    await updateCart({ region_id: region.id })
-    revalidateTag("cart")
+  let cartId = getStoredCart()
+
+  if (!cartId) {
+    const region = await getRegion({ country_code })
+  
+    if (!region) {
+      throw new Error(`Region not found for country code: ${country_code}`)
+    }
+    // Create new cart
+    cartId = (await createCart({ region_id: region.id })).id
   }
 
-  revalidateTag("regions")
-  revalidateTag("products")
+  if (!cartId) {
+    throw new Error("Error retrieving or creating cart")
+  }
 
-  redirect(`/${countryCode}${currentPath}`)
+  const response = await sdk.store.cart.createLineItem(
+    cartId,
+    {
+      variant_id,
+      quantity,
+    },
+    {
+      fields,
+    }
+  )
+
+  return response.cart
+}
+
+/**
+ * Updates the quantity of a line item in the cart.
+ * 
+ * @param line_id - The line item ID to update
+ * @param quantity - The new quantity for the line item
+ * @param fields - Optional fields to include in the response
+ * @returns Promise that resolves to the updated cart
+ * 
+ * @example
+ * ```typescript
+ * // Update line item quantity
+ * const updatedCart = await updateLineItem({
+ *   line_id: 'li_123',
+ *   quantity: 3,
+ *   fields: '*items, *items.variant, shipping_methods'
+ * });
+ * 
+ * // Update with minimal response
+ * const cart = await updateLineItem({
+ *   line_id: 'li_456',
+ *   quantity: 0, // This effectively removes the item
+ *   fields: 'id, total'
+ * });
+ * ```
+ */
+export const updateLineItem = async ({
+  line_id,
+  quantity,
+  fields = DEFAULT_CART_FIELDS,
+}: {
+  line_id: string;
+  quantity: number;
+  fields?: string;
+}): Promise<HttpTypes.StoreCart> => {
+  const cartId = getStoredCart()
+
+  if (!cartId) {
+    throw new Error("No cart found")
+  }
+
+  const { cart } = await sdk.store.cart.updateLineItem(
+    cartId,
+    line_id,
+    { quantity },
+    {
+      fields,
+    }
+  )
+  return cart
+}
+
+/**
+ * Deletes a line item from the cart.
+ * 
+ * @param line_id - The line item ID to delete
+ * @param fields - Optional fields
+ * @returns Promise that resolves when the line item is deleted
+ * 
+ * @example
+ * ```typescript
+ * // Delete a line item
+ * await deleteLineItem({ line_id: 'li_123' });
+ * 
+ * // Delete multiple line items (call multiple times)
+ * await Promise.all([
+ *   deleteLineItem({ line_id: 'li_123' }),
+ *   deleteLineItem({ line_id: 'li_456' })
+ * ]);
+ * ```
+ */
+export const deleteLineItem = async ({
+  line_id,
+  // fields = DEFAULT_CART_FIELDS,
+}: {
+  line_id: string;
+  fields?: string;
+}): Promise<void> => {
+  const cartId = getStoredCart()
+
+  if (!cartId) {
+    throw new Error("No cart found")
+  }
+
+  // TODO pass fields when supported
+  await sdk.store.cart.deleteLineItem(cartId, line_id)
+}
+
+/**
+ * Applies a promotion code to the current cart.
+ * 
+ * @param code - The promotion code to apply
+ * @returns Promise that resolves to the updated cart with applied promotion
+ * 
+ * @example
+ * ```typescript
+ * // Apply a discount code
+ * const cartWithDiscount = await applyPromoCode({ 
+ *   code: 'SAVE20' 
+ * });
+ * 
+ * // Apply a free shipping code
+ * const cartWithFreeShipping = await applyPromoCode({ 
+ *   code: 'FREESHIP' 
+ * });
+ * ```
+ */
+export const applyPromoCode = async ({
+  code,
+}: {
+  code: string;
+}): Promise<HttpTypes.StoreCart> => {
+  const cartId = getStoredCart()
+
+  if (!cartId) {
+    throw new Error("No cart found")
+  }
+
+  const { cart } = await sendPostRequest<{ cart: HttpTypes.StoreCart }>(
+    `/store/carts/${cartId}/promotions`, 
+    {
+      body: {
+        promo_codes: [code],
+      },
+    },
+  )
+
+  return cart
+}
+
+/**
+ * Removes a promotion code from the current cart.
+ * 
+ * @param code - The promotion code to remove
+ * @returns Promise that resolves to the updated cart without the promotion
+ * 
+ * @example
+ * ```typescript
+ * // Remove a discount code
+ * const cartWithoutDiscount = await removePromoCode({ 
+ *   code: 'SAVE20' 
+ * });
+ * 
+ * // Remove a free shipping code
+ * const cartWithoutFreeShipping = await removePromoCode({ 
+ *   code: 'FREESHIP' 
+ * });
+ * ```
+ */
+export const removePromoCode = async ({
+  code,
+}: {
+  code: string;
+}): Promise<HttpTypes.StoreCart> => {
+  const cartId = getStoredCart()
+
+  if (!cartId) {
+    throw new Error("No cart found")
+  }
+
+  const { cart } = await sendDeleteRequest<
+    { cart: HttpTypes.StoreCart }
+  >(`/store/carts/${cartId}/promotions`, {
+    body: {
+      promo_codes: [code],
+    },
+  })
+
+  return cart
 }
